# specs/002-coach-availability-impersonation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PracticePerfect — Coach Invitations, Availability & Impersonation
  version: 1.3.0
  description: |
    Contract for feature 002-coach-availability-impersonation. It is version **1.3.0** of the same
    `/api/v1` contract feature 001 defined at 1.2.0, and it is **purely additive**: every operation
    below is new, and the only changes to existing response models are new optional fields
    (`CurrentUser.impersonation`, `CurrentUser.impersonation_ended`,
    `TrainerPlayerSummary.availability`, `TrainerPlayerSummary.availability_updated_at`). No
    operation is removed, and no existing field changes type or meaning, so the frontend compiled
    against 1.2.0 keeps working until it chooses to read the new fields (research.md R2-22).

    This file documents **only what 1.3.0 adds**. Operations defined in
    `specs/001-user-roles-admin/contracts/openapi.yaml` are unchanged and not repeated here, with
    the single exception of the two schema extensions above, which are given in full so the
    frontend types have one authoritative source.

    Three groups of operations:

    * **Coach invitations** (`coach-invitations` tag, plus `/trainer/coach-invitations` and
      `/trainer/coaches`) — a trainer invites a coach, tracks the invitation, and manages the
      roster; the invited person previews, registers, or accepts. US-01.08, FR-001 – FR-023.
    * **Availability** (`availability` tag) — a coach's own week, a player profile's week, and the
...

...

...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /trainer/coach-invitations:
    get:
      tags: [trainer]
      summary: List the coach invitations this trainer has issued
      operationId: listCoachInvitations
      x-requirements: [FR-004, FR-009]
      description: |
        Trainer only, scoped to the caller's own invitations with no parameter that could widen it.
        Rows in the `superseded` state are never returned: a resend leaves one invitation per
        address, not two (FR-005). `state` is the *presented* state, which folds in expiry and the
        block annotation by the precedence in data-model.md §101.1 — a client never computes it.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: state
          in: query
          required: false
          description: Optional filter on the presented state.
          schema:
            type: string
            enum: [awaiting, accepted, expired, revoked, blocked]
      responses:
        '200':
          description: One page of invitations, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoachInvitationPage' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [trainer]
      summary: Invite a coach by email address
      operationId: issueCoachInvitation
      x-requirements: [FR-001, FR-002, FR-003, FR-007, FR-008, FR-010, FR-023]
      description: |
        Issues a single-use invitation that stops working seven days after issue, and mails it to
        the address with the trainer's name and brand and the personal message when one was given.

        The response never reveals whether the address already holds an account (FR-008): the
        outcome is identical either way, and every conflict that depends on the invited person's
        account surfaces later, to that person, at acceptance.

        Refused with `409 coach_invitation_pending` when this trainer already has an invitation to
        the same address that is awaiting a response and not yet expired; the error body names the
        existing invitation so the client can offer resend or revoke (FR-007). Refused with
        `403 account_not_active` when the trainer's own account is not Active (FR-010).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CoachInvitationCreate' }
      responses:
        '201':
          description: The invitation, as the trainer's list will show it.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoachInvitation' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: An invitation to this address is already awaiting a response (FR-007).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoachInvitationConflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /coach-invitations/{token}:
    get:
      tags: [coach-invitations]
      summary: Preview a coach invitation
      operationId: previewCoachInvitation
      security: []
      x-requirements: [FR-011, FR-012, FR-013]
      description: |
        Public and unauthenticated: possession of the mailed token is the only credential, and it
        was sent to the invited address. Per-origin throttled on the same counter and threshold the
        player join-link lookup uses (research.md R2-05); a refused lookup counts against it.

        `account_exists` tells the client whether to offer registration or sign-in. This is not the
        enumeration FR-008 forbids — that rule protects the trainer from learning whether an address
        is registered, and this response is gated on a 256-bit token mailed to that address.

        Every unusable invitation — unknown token, spent, revoked, superseded, expired, or an
        inviting trainer who is no longer Active — returns the same `404 invitation_link_invalid`
        body, so a stranger cannot learn which condition applied.
      parameters:
        - $ref: '#/components/parameters/CoachInvitationToken'
      responses:
        '200':
          description: What the invited person needs to decide, and nothing else.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoachInvitationPreview' }
        '404': { $ref: '#/components/responses/InvitationLinkInvalid' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /coach-invitations/{token}/register:
    post:
      tags: [coach-invitations]
      summary: Establish a coach account through an invitation and join the roster
      operationId: registerThroughCoachInvitation
      security: []
      x-requirements: [FR-011, FR-013, FR-017, FR-018, FR-023]
      description: |
        For an invited person who holds no account. The email address is **not** in the request
        body: it is taken from the invitation, which is what makes FR-013's address binding
        unbreakable rather than checked. The role is Coach and the trainer is the inviting trainer;
        neither is a parameter, because the invitation already determines both.

        On success the invitation is spent, the coach is on the roster with `joined_at` set, and the
        response sets the session cookie — the coach lands signed in, in a portal already carrying
        the trainer's brand.

        `409 email_already_registered` means an account exists at the invited address; the client
        offers sign-in and `POST .../accept` instead.
      parameters:
        - $ref: '#/components/parameters/CoachInvitationToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CoachRegistrationRequest' }
      responses:
        '201':
          description: The account was created, signed in, and placed on the roster.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoachJoinResult' }
          headers:
            Set-Cookie:
              description: The session cookie for the new coach account.
              schema: { type: string }
        '404': { $ref: '#/components/responses/InvitationLinkInvalid' }
        '409': { $ref: '#/components/responses/Conflict' }
        '410':
          description: The invitation was used while this request was in flight (FR-018).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /coach-invitations/{token}/accept:
    post:
      tags: [coach-invitations]
      summary: Accept a coach invitation with an existing account
      operationId: acceptCoachInvitation
      x-requirements: [FR-012, FR-013, FR-014, FR-015, FR-016, FR-017, FR-018, FR-019, FR-023]
      description: |
        For a signed-in account. Four refusals, each with its own code so the client can say
        something true:

        * `403 coach_invitation_address_mismatch` — the account's address is not the invited one
          (FR-013). The message names the address the invitation was issued for, which the holder of
          the mail already knows.
        * `403 role_cannot_accept` — the account's role is not Coach (FR-014). No role is ever
          changed by following a link.
        * `409 coach_already_assigned` — this coach already works with a trainer (FR-015). The
          message and body name **no** trainer. The invitation is *not* spent and stays usable until
          it expires, so a coach who later leaves that trainer can still accept it.
        * `200 already_on_this_roster` — this coach is already on the inviting trainer's roster
          (FR-016). Not an error: nothing changes and the response says so.

        The two `403`/`409` refusals annotate the invitation as blocked, which is what the trainer
        sees (FR-019); the annotation clears on a later successful acceptance.
      parameters:
        - $ref: '#/components/parameters/CoachInvitationToken'
      responses:
        '200':
          description: |
            The coach is on the roster. `outcome` distinguishes a fresh acceptance from FR-016's
            no-op.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoachJoinResult' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403':
          description: Address mismatch (FR-013) or a role that cannot accept (FR-014).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/InvitationLinkInvalid' }
        '409':
          description: This coach already works with a trainer (FR-015). Names no trainer.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }

  # ==========================================================================
  # Availability (FR-024 – FR-039)
  # ==========================================================================
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pp_session
      description: |
        Unchanged from feature 001. While an impersonation is live the same cookie resolves to the
        impersonated account; no second credential is issued (research.md R2-14).
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
    CoachInvitationToken:
      name: token
      in: path
      required: true
      description: |
        The raw token from the emailed URL. Only its SHA-256 is stored (research.md R2-02), so this
        value cannot be recovered from the database or read back by the trainer.
      schema: { type: string, minLength: 32, maxLength: 64 }
...
  responses:
    Unauthenticated:
      description: No usable session cookie.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: The caller's role or ownership does not permit this action.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflict:
      description: The request conflicts with existing state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    InvitationLinkInvalid:
      description: |
        The invitation cannot be used. One body for every reason — unknown, spent, revoked,
        superseded, expired, inviting trainer no longer Active.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many invitation lookups from this origin.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: One or more fields are invalid; each offending field is identified.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorResponse' }
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: true
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    ValidationErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [message, fields]
          properties:
            message: { type: string }
            fields:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [field, message]
                properties:
                  field: { type: string }
                  message: { type: string }

    # --- Coach invitations ---
    CoachInvitationCreate:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
          description: The invited address. Lower-cased on receipt.
        invitee_name:
          type: [string, 'null']
          minLength: 1
          maxLength: 200
          description: |
            Optional. `null` when not given — never `""`, which is a 422 (Principle VI).
        message:
          type: [string, 'null']
          minLength: 1
          maxLength: 2000
          description: Optional personal message, carried into the invitation email.
    CoachInvitationState:
      type: string
      enum: [awaiting, accepted, expired, revoked, blocked]
      description: |
        The **presented** state (data-model.md §101.1). `superseded` is not in this enum because a
        superseded invitation is never returned. `expired` and `blocked` are derived, not stored.
    CoachInvitation:
      type: object
      additionalProperties: false
      required:
        [id, invited_email, invitee_name, message, state, issued_at, expires_at, accepted_at,
         revoked_at, blocked_reason, coach]
      properties:
        id: { type: string, format: uuid }
        invited_email: { type: string, format: email }
        invitee_name: { type: [string, 'null'] }
        message: { type: [string, 'null'] }
        state: { $ref: '#/components/schemas/CoachInvitationState' }
        issued_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        accepted_at: { type: [string, 'null'], format: date-time }
        revoked_at: { type: [string, 'null'], format: date-time }
        blocked_reason:
          type: [string, 'null']
          enum: [role_not_coach, already_assigned, null]
          description: |
            Why an acceptance was refused (FR-019). `already_assigned` never carries or implies the
            identity of the trainer the coach already works with (FR-015).
        coach:
          oneOf:
            - $ref: '#/components/schemas/CoachSummary'
            - type: 'null'
          description: The coach who accepted, once one has. `null` in every other state.
    CoachInvitationPage:
      type: object
      additionalProperties: false
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CoachInvitation' }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
    CoachInvitationConflict:
      type: object
      additionalProperties: false
      required: [error]
      description: FR-007's 409 — the existing invitation is named so the client can act on it.
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, invitation]
          properties:
            code: { type: string, enum: [coach_invitation_pending] }
            message: { type: string }
            invitation: { $ref: '#/components/schemas/CoachInvitation' }
    CoachInvitationPreview:
      type: object
      additionalProperties: false
      required: [invited_email, invitee_name, message, expires_at, account_exists, trainer]
      properties:
        invited_email:
          type: string
          format: email
          description: |
            Returned in full, not masked: the reader received the mail at this address, and FR-013
            requires them to be told which address the invitation binds to.
        invitee_name: { type: [string, 'null'] }
        message: { type: [string, 'null'] }
        expires_at: { type: string, format: date-time }
        account_exists:
          type: boolean
          description: Whether an account already exists at the invited address (register vs sign in).
        trainer:
          type: object
          additionalProperties: false
          required: [business_name, portal_branding]
          description: |
            Enough to show whose organization this is, and to brand the page before sign-in — the
            same public branding the player join page already renders. The trainer's public
            identity only: no address, no contact detail, no counts.
          properties:
            business_name: { type: string }
            portal_branding:
              type: object
              additionalProperties: true
              description: As defined by feature 001's `PortalBranding`.
    CoachRegistrationRequest:
      type: object
      additionalProperties: false
      required: [first_name, last_name, password]
      description: |
        No `email` and no `role`: both come from the invitation (FR-011, FR-013). No `trainer_id`
        either — the invitation names the trainer.
      properties:
        first_name: { type: string, minLength: 1, maxLength: 100 }
        last_name: { type: string, minLength: 1, maxLength: 100 }
        password:
          type: string
          minLength: 12
          description: Subject to the same password policy every other account is.
        phone:
          type: [string, 'null']
          minLength: 1
          maxLength: 32
        bio:
          type: [string, 'null']
          minLength: 1
          maxLength: 2000
        credentials:
          type: [string, 'null']
          minLength: 1
          maxLength: 1000
        certifications:
          type: [string, 'null']
          minLength: 1
          maxLength: 1000
    CoachJoinResult:
      type: object
      additionalProperties: false
      required: [outcome, trainer_business_name, joined_at]
      properties:
        outcome:
          type: string
          enum: [joined, already_on_this_roster]
          description: FR-016's no-op is reported as an outcome, not as an error.
        trainer_business_name: { type: string }
        joined_at: { type: string, format: date-time }
    CoachSummary:
      type: object
      additionalProperties: false
      required: [user_id, first_name, last_name, email, status, photo_url]
      properties:
        user_id: { type: string, format: uuid }
        first_name: { type: string }
        last_name: { type: string }
        email: { type: string, format: email }
        status: { type: string, enum: [active, inactive, deleted] }
        photo_url: { type: [string, 'null'] }
...